+++ src/main/java/Domain/Repositories/InterfaceShoppingCartRepository.java
@@ -9,7 +9,7 @@ import Domain.Entities.ShoppingCart;
 import Domain.Entities.User;
 
 @NoRepositoryBean
-public interface InterfaceShoppingCartRepository extends JpaRepository<ShoppingCart, Long> {
+public interface InterfaceShoppingCartRepository extends JpaRepository<ShoppingCart, Integer> {
 
     @Query("SELECT CASE WHEN COUNT(c) > 0 THEN true ELSE false END FROM ShoppingCart c WHERE c._username = ?1")
     ShoppingCart getCartByUsername(String username);
